# openapi.yml
openapi: 3.0.0

info:
  title: "OAuthZEN Authorization API"
  version: "1.0"
  description: "A simple API to demonstrate OpenAPI 3.1 features"

servers:
  - url: "http://localhost:8080/access/v1"

paths:
  /access/v1/evaluation:
    post:
      summary: "Evaluate a user"
      description: "Evaluate a user based on their profile"
      operationId: evaluateAccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subject:
                  type: object
                  properties:
                    type:
                      type: string
                    id:
                      type: string
                    properties:
                      type: object
                      additionalProperties: true
                  required:
                    - type
                    - id
                resource:
                  type: object
                  properties:
                    type:
                      type: string
                    id:
                      type: string
                    properties:
                      type: object
                      properties:
                        ownerID:
                          type: string
                      additionalProperties: true
                  required:
                    - type
                    - id
                action:
                  type: object
                  properties:
                    name:
                      type: string
                    properties:
                      type: object
                      additionalProperties: true
                  required:
                    - name
                constext:
                  type: object
                  additionalProperties: true
              required:
                - subject
                - resource
                - action
              example:
                subject:
                  type: "user"
                  id: "alice@acmecorp.com"
                resource:
                  type: "account"
                  id: "123"
                action:
                  name: "can_read"
                  properties:
                    method: "GET"
                context:
                  time: "2025-01-01T00:00:00Z"
      responses:
        '200':
          description: "Evaluation result"
          content:
            application/json:
              schema:
                type: object
                properties:
                  decision:
                    type: boolean
                required:
                  - decision